IGAPP-114: Move LayoutedScrollView to containers
