+++ native/src/routes/offers/containers/OffersContainer.js
@@ -24,7 +24,7 @@ import {
 } from '../../../modules/app/components/NavigationTypes'
 import type { StoreActionType } from '../../../modules/app/StoreActionType'
 import type { Dispatch } from 'redux'
-import LayoutedScrollView from '../../../modules/common/components/LayoutedScrollView'
+import LayoutedScrollView from '../../../modules/common/containers/LayoutedScrollView'
 import LayoutContainer from '../../../modules/layout/containers/LayoutContainer'
 import { cityContentUrl } from '../../../modules/common/url'
 import openExternalUrl from '../../../modules/common/openExternalUrl'
